+++ src/main/java/com/volomak/movieland/dao/postgres/GenreDaoImpl.java
@@ -1,17 +1,14 @@
-package com.volomak.movieland.dao.jdbc;
+package com.volomak.movieland.dao.postgres;
 
 import com.volomak.movieland.dao.GenreDao;
-import com.volomak.movieland.dao.jdbc.mapper.GenreRowMapper;
+import com.volomak.movieland.dao.postgres.mapper.GenreRowMapper;
 import com.volomak.movieland.entity.Genre;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Service;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 
 @Service
