+++ src/main/java/com/volomak/movieland/dao/postgres/mapper/ReviewRowMapper.java
@@ -1,4 +1,4 @@
-package com.volomak.movieland.dao.jdbc.mapper;
+package com.volomak.movieland.dao.postgres.mapper;
 
 import com.volomak.movieland.entity.Review;
 import org.springframework.jdbc.core.RowMapper;
